Refactoring

+++ src/View/Inputs/DefinitionManager.cpp
@@ -45,7 +45,7 @@ auto DefinitionManager::AddDefinitionWidgets() -> void
         ProcessType::GBM,
     };
     auto* processes = new QComboBox(this);
-    for (int i = 0; i < processTypes.size(); ++i) {
+    for (u_int8_t i = 0; i < processTypes.size(); ++i) {
         processes->insertItem(i, QString::fromStdString(std::string(ProcessData::GetAcronym(processTypes[i]))));
         processes->setItemData(i, QString::fromStdString(std::string(ProcessData::GetName(processTypes[i]))), Qt::ToolTipRole);
     }
@@ -66,7 +66,7 @@ auto DefinitionManager::AddDefinitionWidgets() -> void
         qobject_cast<QComboBox*>(m_widgets[DefinitionWidget::PROCESS]),
         QOverload<int>::of(&QComboBox::currentIndexChanged),
         this,
-        [this, processTypes](int newIndex) {
+        [this, processTypes](u_int8_t newIndex) {
             Parent()->OnProcessTypeModified(processTypes[newIndex]);
             UpdateWidgetProperties(processTypes[newIndex]);
         }
